- Applying defined naming rules for cps projects to support QA Type system.
