+++ server/src/test/scala/za/co/absa/atum/server/api/database/runs/functions/CreateOrUpdateAdditionalDataIntegrationTest.scala
@@ -18,7 +18,7 @@ package za.co.absa.atum.server.api.database.runs.functions
 
 import org.junit.runner.RunWith
 import za.co.absa.atum.model.dto.{AdditionalDataSubmitDTO, PartitionDTO}
-import za.co.absa.atum.server.ConfigProviderSpec
+import za.co.absa.atum.server.ConfigProviderTest
 import za.co.absa.atum.server.api.TestTransactorProvider
 import za.co.absa.atum.server.api.database.PostgresDatabaseProvider
 import za.co.absa.fadb.exceptions.DataNotFoundException
@@ -28,7 +28,7 @@ import zio.test._
 import zio.test.junit.ZTestJUnitRunner
 
 @RunWith(classOf[ZTestJUnitRunner])
-class CreateOrUpdateAdditionalDataIntegrationSpec extends ConfigProviderSpec {
+class CreateOrUpdateAdditionalDataIntegrationTest extends ConfigProviderTest {
 
   override def spec: Spec[TestEnvironment with Scope, Any] = {
