+++ server/src/test/scala/za/co/absa/atum/server/api/service/CheckpointServiceIntegrationTest.scala
@@ -14,52 +14,55 @@
  * limitations under the License.
  */
 
-package za.co.absa.atum.server.api.repository
+package za.co.absa.atum.server.api.service
 
 import org.junit.runner.RunWith
 import org.mockito.Mockito.{mock, when}
-import za.co.absa.atum.server.api.database.runs.functions.WriteCheckpoint
-import za.co.absa.atum.server.api.exception.DatabaseError
 import za.co.absa.atum.server.api.TestData
+import za.co.absa.atum.server.api.exception.{DatabaseError, ServiceError}
+import za.co.absa.atum.server.api.repository.CheckpointRepository
 import za.co.absa.fadb.exceptions.ErrorInDataException
 import za.co.absa.fadb.status.FunctionStatus
-import zio._
 import zio.test.Assertion.failsWithA
 import zio.test._
+import zio._
 import zio.test.junit.ZTestJUnitRunner
 
 @RunWith(classOf[ZTestJUnitRunner])
-class CheckpointRepositoryIntegrationSpec extends ZIOSpecDefault with TestData {
+class CheckpointServiceIntegrationTest extends ZIOSpecDefault with TestData {
 
-  private val writeCheckpointMock: WriteCheckpoint = mock(classOf[WriteCheckpoint])
+  private val checkpointRepositoryMock = mock(classOf[CheckpointRepository])
 
-  when(writeCheckpointMock.apply(checkpointDTO1)).thenReturn(ZIO.right(()))
-  when(writeCheckpointMock.apply(checkpointDTO2))
+  when(checkpointRepositoryMock.writeCheckpoint(checkpointDTO1)).thenReturn(ZIO.right(()))
+  when(checkpointRepositoryMock.writeCheckpoint(checkpointDTO2))
     .thenReturn(ZIO.left(ErrorInDataException(FunctionStatus(50, "error in data"))))
-  when(writeCheckpointMock.apply(checkpointDTO3))
-    .thenReturn(ZIO.fail(new Exception("boom!")))
+  when(checkpointRepositoryMock.writeCheckpoint(checkpointDTO3))
+    .thenReturn(ZIO.fail(DatabaseError("boom!")))
 
-  private val writeCheckpointMockLayer = ZLayer.succeed(writeCheckpointMock)
+  private val checkpointRepositoryMockLayer = ZLayer.succeed(checkpointRepositoryMock)
 
   override def spec: Spec[TestEnvironment with Scope, Any] = {
 
-    suite("CheckpointRepositoryIntegrationSuite")(
-      suite("WriteCheckpointSuite")(
+    suite("CheckpointServiceIntegrationSuite")(
+      suite("SaveCheckpointSuite")(
         test("Returns expected Right with Unit") {
           for {
-            result <- CheckpointRepository.writeCheckpoint(checkpointDTO1)
+            result <- CheckpointService.saveCheckpoint(checkpointDTO1)
           } yield assertTrue(result.isRight)
         },
         test("Returns expected Left with StatusException") {
           for {
-            result <- CheckpointRepository.writeCheckpoint(checkpointDTO2)
+            result <- CheckpointService.saveCheckpoint(checkpointDTO2)
           } yield assertTrue(result.isLeft)
         },
-        test("Returns expected DatabaseError") {
-          assertZIO(CheckpointRepository.writeCheckpoint(checkpointDTO3).exit)(failsWithA[DatabaseError])
+        test("Returns expected ServiceError") {
+          assertZIO(CheckpointService.saveCheckpoint(checkpointDTO3).exit)(failsWithA[ServiceError])
         }
       )
-    ).provide(CheckpointRepositoryImpl.layer, writeCheckpointMockLayer)
+    ).provide(
+      CheckpointServiceImpl.layer,
+      checkpointRepositoryMockLayer
+    )
 
   }
